add validate to managecar page

+++ src/components/admin/ManageCarButton.tsx
@@ -93,6 +93,10 @@ const ManageCarButton: React.FC<props> = ({
   const [inputGear, setInputGear] = React.useState("");
   const [typeCar, settypeCar] = React.useState("");
   const [typeGear, setTypeGear] = React.useState("");
+  const [errorName, setErrorName] = useState("");
+  const [errorSeats, setErrorSeats] = useState("");
+  const [errorDoors, setErrorDoors] = useState("");
+  const [errorCost, setErrorCost] = useState("");
   const typeId = new Map<string, string>([
     ["Sedan", "1"],
     ["Van", "2"],
@@ -117,13 +121,59 @@ const ManageCarButton: React.FC<props> = ({
     preview: img,
     raw: img,
     typeID: options[type - 1],
-    gear_type: gear == "A" ? "Auto":"Manual",
+    gear_type: gear == "A" ? "Auto" : "Manual",
     seats: seats,
     doors: doors,
   });
   // เปลี่ยนแปลงค่า
   const handleChange =
     (prop: keyof State) => (event: React.ChangeEvent<HTMLInputElement>) => {
+      const { value } = event.target;
+      if (prop == "carName") {
+        if (value.length == 0) {
+          setErrorName("Car name field should not empty! ");
+        } else if (value == " ") {
+          setErrorName("first character shound not be space!");
+        } else if (/^[a-zA-Z0-9ก-ฮ]+\s+[a-zA-Z0-9]+\s+\d{4}$/.test(value)) {
+          setErrorName("");
+        } else if (/^[a-zA-Z0-9ก-ฮ]+\s+[a-zA-Z0-9]+\s+\D*$/.test(value)) {
+          setErrorName("Invalid name field!");
+        }
+        setValues({ ...values, [prop]: value });
+      } else if (prop == "seats") {
+        if (value.length == 0) {
+          setErrorSeats("Seats field should not empty!");
+        } else if (event.target.value[0] == " ") {
+          setErrorSeats("first character shound not space!");
+        } else if (/^\d{1}$/.test(value)) {
+          setErrorSeats("");
+        } else {
+          setErrorSeats("Invalid seats field!");
+        }
+        setValues({ ...values, [prop]: value });
+      } else if (prop == "doors") {
+        if (value.length == 0) {
+          setErrorDoors("Doors field should not empty!");
+        } else if (event.target.value[0] == " ") {
+          setErrorDoors("first character shound not space!");
+        } else if (/^\d{1}$/.test(value)) {
+          setErrorDoors("");
+        } else {
+          setErrorDoors("Invalid doors field!");
+        }
+        setValues({ ...values, [prop]: value });
+      } else if (prop == "price") {
+        if (value.length == 0) {
+          setErrorCost("Price field should not empty!");
+        } else if (event.target.value[0] == " ") {
+          setErrorCost("first character shound not space!");
+        } else if (/^\d+$/.test(value)) {
+          setErrorCost("");
+        } else {
+          setErrorCost("Invalid cost field!");
+        }
+        setValues({ ...values, [prop]: value });
+      }
       setValues({ ...values, [prop]: event.target.value });
     };
   //เปลี่ยนรูป
@@ -133,7 +183,7 @@ const ManageCarButton: React.FC<props> = ({
         carName: values.carName,
         carID: values.carID,
         price: values.price,
-        preview: values.preview,//URL.createObjectURL(e.target.files[0]),
+        preview: values.preview, //URL.createObjectURL(e.target.files[0]),
         raw: values.raw,
         typeID: values.typeID,
         gear_type: values.gear_type,
@@ -167,39 +217,42 @@ const ManageCarButton: React.FC<props> = ({
       localStorage.getItem("admin") ?? ' { "token": "" }'
     ).token;
     console.log("token", token);
-
-    axios({
-      method: "PUT",
-      url: `http://localhost:3001/vehicle/edit/${values.carID}`,
-      data: {
-        name: values.carName,
-        vehicle_id: id,
-        cost: values.price,
-        type_id: typeId.get(values.typeID)!,
-        vehicle_img: values.raw,
-        seats: values.seats,
-        doors: values.doors,
-        gear_type: values.gear_type == "Auto" ? "A" : "M",
-      },
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    })
-      .then((response) => {
-        return response.data;
+    if (!(!!errorCost || !!errorDoors || !!errorName || !!errorSeats)) {
+      axios({
+        method: "PUT",
+        url: `http://localhost:3001/vehicle/edit/${values.carID}`,
+        data: {
+          name: values.carName,
+          vehicle_id: id,
+          cost: values.price,
+          type_id: typeId.get(values.typeID)!,
+          vehicle_img: values.raw,
+          seats: values.seats,
+          doors: values.doors,
+          gear_type: values.gear_type == "Auto" ? "A" : "M",
+        },
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
       })
-      .then((data) => {
-        console.log(data);
-      })
-      .catch((error) => {
-        if (error.response.status == "401") {
-          localStorage.clear();
-          setonLoginadmin(false);
-          alert("กรุณาเข้าสู่ระบบใหม่อีกครั้ง");
-          navigate("/Admin");
-        }
-      })
-      .then(handleClose2);
+        .then((response) => {
+          return response.data;
+        })
+        .then((data) => {
+          console.log(data);
+        })
+        .catch((error) => {
+          if (error.response.status == "401") {
+            localStorage.clear();
+            setonLoginadmin(false);
+            alert("กรุณาเข้าสู่ระบบใหม่อีกครั้ง");
+            navigate("/Admin");
+          }
+        })
+        .then(handleClose2);
+    } else {
+      alert("กรุณาตรวจสอบข้อมูล");
+    }
   };
   // เมื่อกดปุ่มลบ
   const handleDelete = (e: React.MouseEvent) => {
@@ -244,7 +297,7 @@ const ManageCarButton: React.FC<props> = ({
       preview: "",
       raw: "",
       typeID: options[type - 1],
-      gear_type: gear == "A" ? "Auto":"Manual",
+      gear_type: gear == "A" ? "Auto" : "Manual",
       seats: seats,
       doors: doors,
     });
@@ -367,13 +420,17 @@ const ManageCarButton: React.FC<props> = ({
                   autoComplete="off"
                 >
                   <TextField
+                    error={!!errorName}
+                    helperText={errorName}
                     id="carName"
-                    label="carName"
+                    label="Vehicle name"
                     variant="outlined"
                     value={values.carName}
                     onChange={handleChange("carName")}
                   />
                   <TextField
+                    error={!!errorDoors}
+                    helperText={errorDoors}
                     id="doors"
                     label="doors"
                     variant="outlined"
@@ -381,6 +438,8 @@ const ManageCarButton: React.FC<props> = ({
                     onChange={handleChange("doors")}
                   />
                   <TextField
+                    error={!!errorSeats}
+                    helperText={errorSeats}
                     id="seats"
                     label="seats"
                     variant="outlined"
@@ -389,6 +448,8 @@ const ManageCarButton: React.FC<props> = ({
                   />
 
                   <TextField
+                    error={!!errorCost}
+                    helperText={errorCost}
                     id="price"
                     label="price"
                     variant="outlined"
